requirements.txt
fastapi
uvicorn
pyTelegramBotAPI
google-generativeai
psycopg2-binary
python-dotenv
redis